Fix some compiler warnings.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@920339 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/activemq/commands/ActiveMQBytesMessage.cpp
@@ -627,7 +627,7 @@ void ActiveMQBytesMessage::storeContent() {
                 DataOutputStream doBuffer( &buffer );
 
                 // Start by writing the length of the written data before compression.
-                doBuffer.writeInt( this->length );
+                doBuffer.writeInt( (int)this->length );
 
                 // Now write the Compressed bytes.
                 this->bytesOut->writeTo( &doBuffer );
